Actually made custom map object
