+++ app/src/main/java/com/chrsrck/quakemap/ui/EarthquakeMapFragment.kt
@@ -63,7 +63,7 @@ class EarthquakeMapFragment : Fragment(), OnMapReadyCallback {
         with(googleMap) {
 //            moveCamera(com.google.android.gms.maps.CameraUpdateFactory.newLatLngZoom(SYDNEY, ZOOM_LEVEL))
 //            addMarker(com.google.android.gms.maps.model.MarkerOptions().position(SYDNEY))
-            googleMap.uiSettings.isMapToolbarEnabled = false
+            val quakeMap = EarthquakeMap(googleMap)
 
             val mapObserver = Observer<Boolean> { darkMode ->
                 // Update the UI, in this case, a TextView.
@@ -76,21 +76,7 @@ class EarthquakeMapFragment : Fragment(), OnMapReadyCallback {
             }
 
             val quakeLiveDataObserver = Observer<HashMap<String, Earthquake>> { quakeHashMap ->
-                //googleMap.clear()
-                quakeHashMap?.values?.map { earthquake: Earthquake ->
-                    val options = MarkerOptions()
-
-                    when(earthquake.magnitude) {
-                        in 0..3 -> options.icon(BitmapDescriptorFactory.defaultMarker(BitmapDescriptorFactory.HUE_GREEN))
-                        in 3..6 -> options.icon(BitmapDescriptorFactory.defaultMarker(BitmapDescriptorFactory.HUE_YELLOW))
-                        in 6..9 -> options.icon(BitmapDescriptorFactory.defaultMarker(BitmapDescriptorFactory.HUE_ORANGE))
-                        else -> { options.icon(BitmapDescriptorFactory.defaultMarker(BitmapDescriptorFactory.HUE_GREEN))}
-                    }
-
-                    googleMap.addMarker(options
-                            .title(earthquake.title)
-                            .position(LatLng(earthquake.latitude, earthquake.longitude)))
-                }
+                quakeMap.refreshQuakes(quakeHashMap)
             }
             viewModel.darkMode.observe(frag, mapObserver)
             viewModel.dataSource.hashMap.observe(frag, quakeLiveDataObserver)
